Django REST/pyproject.toml
[project]
name = "djangoprojectposts"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "crispy-bootstrap5>=2025.6",
    "django>=5.2.4",
    "django-crispy-forms>=2.4",
    "django-debug-toolbar>=6.0.0",
    "django-filter>=25.1",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt[crypto]>=5.5.1",
    "djoser>=2.3.3",
    "faker>=37.5.3",
    "markdown>=3.8.2",
    "pillow>=11.3.0",
    "psycopg2>=2.9.10",
    "redis>=6.4.0",
]

[dependency-groups]
dev = [
    "coverage>=7.10.6",
    "factory-boy>=3.3.3",
]
